Move the Example Global Overload to a common location.
